.Net: Changes to support .NET9 formatting (#10838)
### Motivation and Context

In sdk9, there is a bug fix to one of the analyzers:
[https://github.com/dotnet/msbuild/issues/11426](https://github.com/dotnet/msbuild/issues/11426).
This causes the code base to fail in the `dotnet format` stage when
built on sdk9.

This PR fixes the file that has been changed this week and adds a
configuration into the build to format using sdk9.0 so that the build
fails if more are introduced.

- [x] The code builds clean without any errors or warnings
- [ ] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [ ] All unit tests pass, and I have added new tests where possible
- [ ] I didn't break anyone :smile:

This relates to #10716 and was previously fixed in #10741